--- packages/local-environment/src/index.ts
@@ -1,6 +1,0 @@
-import { LocalNetwork, RemoteNetwork, Network } from "./network";
-import { Key } from "./key";
-import { Nomad } from "./nomad";
-import * as utils from "./utils";
-
-export { RemoteNetwork, LocalNetwork, Network, Key, Nomad, utils };
