--- packages/local-environment/src/types.ts
@@ -1,11 +1,0 @@
-
-// // TODO: the old deploy (what is currently published to npm) has been deprecated
-// import {
-//   BridgeDeploy,
-//   ExistingBridgeDeploy,
-// } from "@nomad-xyz/deploy/lib/bridge/BridgeDeploy";
-
-// export type Domain = number;
-// export type Address = string;
-// export type HexString = string;
-// export type Deploy = BridgeDeploy | ExistingBridgeDeploy;
